refactor: merge relay and core crates
